requirements.txt
langchain
langchain-huggingface
langchain-openai
langchain-community
yt_dlp
spacy
plotly
psycopg2-binary
dash
piper-tts
serpapi
requests
python-dotenv
mcp
fastmcp
fastapi
uvicorn 